backend/requirements.txt
  
django==4.2
joblib==1.4.2
scikit-learn==1.5.1
# psycopg2==2.8.4
psycopg2-binary==2.9.9
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
django-filter==24.3